requirements.txt
streamlit
trimesh
numpy-stl
numpy
